+++ lib/coupon_uikit.dart
@@ -1,4 +1,5 @@
-library coupon_uikit;
+library;
 
 export 'src/clippers/coupon_clipper.dart';
 export 'src/widgets/coupon_card.dart';
+export 'src/painters/coupon_decoration_painter.dart';